Modify the type of variable idx

+++ libraries/AP_Proximity/AP_Proximity_uSharp3D_MT.cpp
@@ -189,7 +189,7 @@ bool AP_Proximity_uSharp3D_MT::get_reading(void) {
 						avoid_dis = final_dis / 100.0;
 						//
 						for (int j = 0; j < target_num; j++) {
-							uint8_t idx = 6 + 9 * j;
+							uint16_t idx = 6 + 9 * j;
 							//
 							uint16_t dis = buf[idx] * 256 + buf[idx + 1];
 							//
